+++ index.html
@@ -150,6 +150,8 @@
 			crossorigin="anonymous"
 			referrerpolicy="no-referrer"
 		></script>
-		<script src="script.js"></script>
+		<script src="script.js" type="module"></script>
+		<script src="./modules/export.js" type="module"></script>
+		<script src="./modules/theme.js"></script>
 	</body>
 </html>
