+++ modules/isEmptyOrSpaces.js
@@ -1,0 +1,3 @@
+export default function isEmptyOrSpaces(str) {
+    return str === null || str.match(/^ *$/) !== null;
+}
